change starcount to long

+++ app/src/main/java/com/example/chayen/cookingsupporter/FoodListAdapter/FoodDatabaseClass.java
@@ -18,14 +18,14 @@ public class FoodDatabaseClass implements Serializable{
     public String food_name;
     public String food_type;
     public ArrayList<String> ingredient;
-    public int star_count;
+    public Long star_count;
 
     public FoodDatabaseClass(){
 
     }
 
     public FoodDatabaseClass(String author, ArrayList<String> cooking_method, String food_image,
-                             String food_name, String food_type, ArrayList<String> ingredient, int star_count){
+                             String food_name, String food_type, ArrayList<String> ingredient, Long star_count){
         this.author = author;
         this.cooking_method = cooking_method;
         this.food_image = food_image;
@@ -83,11 +83,11 @@ public class FoodDatabaseClass implements Serializable{
         this.ingredient = ingredient;
     }
 
-    public int getStar_count(){
+    public Long getStar_count(){
         return star_count;
     }
 
-    public void setStar_count(int star_count){
+    public void setStar_count(Long star_count){
         this.star_count = star_count;
     }
